requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
drf-spectacular>=0.27
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
PyYAML>=6.0
